requirements.txt
mss==9.0.2
opencv-python==4.9.0.80
pytesseract==0.3.10
Pillow==10.3.0
pyautogui==0.9.54
pyyaml==6.0.1
numpy==1.26.4
dataclasses-json==0.6.4
loguru==0.7.2
